refactor: consolidate create database prep form components
