+++ frontend/src/components/CreateDatabasePrepForm/index.ts
@@ -1,11 +1,3 @@
-import CreateDatabasePrepButtonGroup from "./CreateDatabasePrepButtonGroup.vue";
-import CreateDatabasePrepForm from "./CreateDatabasePrepForm.vue";
 import CreateDatabasePrepPanel from "./CreateDatabasePrepPanel.vue";
 
-export {
-  CreateDatabasePrepForm,
-  CreateDatabasePrepPanel,
-  CreateDatabasePrepButtonGroup,
-};
-
-export default CreateDatabasePrepForm;
+export { CreateDatabasePrepPanel };
